refactor(runtime): isolate PRG native object helpers
